Major bug fix
> Added check for error parameter validity in engine file;

On some servers, after a normal error would pop up on a client,
LuaError2 would run the hook ~5 times with all arguments nil, stacking
up to 13 times precisely. This caused massive error spam on the client
and slowed the server to less than one frame per second.
The actual cause is unknown, and so is why it only happens on some
servers.

The problem sources from, or at least goes through LuaError2.